feat: Add product inclusions field
Introduces an `inclusions` field to the `Product` type to specify what comes in the box. This change includes updating the Product Detail Page to display this information and the Admin Dashboard to allow editing of inclusions. The `sheetsService` is also updated to handle the import and export of this new field.